Test #1 outcome
Test #1 was successful. There was a problem where when I git clone my repository to Digital Ocean, my old resume would show on the site. I moved the location of my new resume and made changes to the html code and now my new resume displays on the site. Will need to look into this problem move in the future.